# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/users":{"post":{"operationId":"UsersController_createUser","summary":"Create the user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"Success"},"400":{"description":"Fail"}},"tags":["users"]},"get":{"operationId":"UsersController_getUsers","summary":"Get all users","parameters":[{"name":"withOrder","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Success"}},"tags":["users"]}},"/api/users/{userId}":{"get":{"operationId":"UsersController_getUser","summary":"Get the user","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}},{"name":"withOrder","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Success"}},"tags":["users"]},"put":{"operationId":"UsersController_updateUser","summary":"Update the user","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"Success"},"400":{"description":"Fail"}},"tags":["users"]},"delete":{"operationId":"UsersController_deleteUser","summary":"Delete the user","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success"}},"tags":["users"]}},"/api/orders":{"post":{"operationId":"OrdersController_createUser","summary":"Create the order","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"200":{"description":"Success"},"400":{"description":"Fail"}},"tags":["orders"]},"get":{"operationId":"OrdersController_getOrders","summary":"Get all orders","parameters":[],"responses":{"200":{"description":"Success"}},"tags":["orders"]}},"/api/orders/{orderId}":{"get":{"operationId":"OrdersController_getUser","summary":"Get the order","parameters":[{"name":"orderId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success"}},"tags":["orders"]},"put":{"operationId":"OrdersController_updateUser","summary":"Update the order","parameters":[{"name":"orderId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrderDto"}}}},"responses":{"200":{"description":"Success"},"400":{"description":"Fail"}},"tags":["orders"]},"delete":{"operationId":"OrdersController_deleteUser","summary":"Delete the order","parameters":[{"name":"orderId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success"}},"tags":["orders"]}}},"info":{"title":"Delivery API","description":"Delivery API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"user@gmail.com"},"username":{"type":"string","example":"user"},"password":{"type":"string","example":"secure"},"address":{"type":"string","example":"ул. Мира, 32, Екатеринбург"},"phoneNumber":{"type":"string","example":"83233654865"}},"required":["email","username","password","address","phoneNumber"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"user@gmail.com"},"username":{"type":"string","example":"user"},"address":{"type":"string","example":"ул. Мира, 32, Екатеринбург"},"phoneNumber":{"type":"string","example":"83233654865"}},"required":["email","username","address","phoneNumber"]},"CreateOrderDto":{"type":"object","properties":{"deliveryAddress":{"type":"string","example":"ул. Мира, 32, Екатеринбург"},"paymentMethod":{"type":"string","example":"card"},"status":{"type":"string","example":"processing"},"userId":{"type":"string"}},"required":["deliveryAddress","paymentMethod","status","userId"]},"UpdateOrderDto":{"type":"object","properties":{"deliveryAddress":{"type":"string","example":"ул. Мира, 32, Екатеринбург"},"paymentMethod":{"type":"string","example":"cash"},"status":{"type":"string","example":"delivered"}},"required":["deliveryAddress","paymentMethod","status"]}}}}